Test: fix name-clash issue
