+++ src/Mod/Test/Gui/CMakeLists.txt
@@ -67,7 +67,7 @@ fc_copy_sources(QtUnitGui "${CMAKE_BINARY_DIR}/Mod/Test" qtunittest.py)
 SET_BIN_DIR(QtUnitGui QtUnitGui /Mod/Test)
 SET_PYTHON_PREFIX_SUFFIX(QtUnitGui)
 
-INSTALL(TARGETS QtUnitGui DESTINATION ${CMAKE_INSTALL_LIBDIR})
+INSTALL(TARGETS QtUnitGui DESTINATION Mod/Test)
 INSTALL(FILES ${TestGuiPy_SRCS} DESTINATION Mod/Test)
 
 SET(TestGuiIcon_SVG
