+++ src/Mod/Test/Init.py
@@ -24,3 +24,6 @@
 #*                                                                         *
 #*   Juergen Riegel 2002                                                   *
 #***************************************************************************/
+
+
+removeFromPath("Test")
